requirements.txt
pygame==2.6.1
google-generativeai>=0.8.0
python-dotenv>=1.0.0